Unfolding macro reweighting years with proper cross section
Restructured macro using the new
handler classes for varous steps. The
macro sums up the raw level spectra
and the integrated luminosities from the various years, using for the spectra
now the absolute counts and not the
downscale-weighted counts. The rest
of the unfolding chain is the same
as in previous versions of the macro.